% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{coefficientMatrixSpheres}
\alias{coefficientMatrixSpheres}
\title{Calculate coefficients (spheres)}
\usage{
coefficientMatrixSpheres(bin)
}
\arguments{
\item{bin}{non-decreasing vector of class limits}
}
\value{
array of coefficients
}
\description{
Matrix of coefficients for Wicksell's corpuscle problem
}
\details{
The function calculates the matrix of coefficients of the
discretized integral equation for Wicksell's corpuscle problem.
}
\references{
Ohser, J. and Muecklich, F. Statistical analysis of microstructures in materials science J. Wiley & Sons, 2000
}
