% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylinder.R
\name{cylinders3d}
\alias{cylinders3d}
\title{Plot fibre system}
\usage{
cylinders3d(S, box, draw.axes = FALSE, draw.box = TRUE, clipping = FALSE,
  ...)
}
\arguments{
\item{S}{a list of cylinders}

\item{box}{simulation box}

\item{draw.axes}{logical: if true, draw the axes}

\item{draw.box}{logical: if true, draw the bounding box}

\item{clipping}{logical: if true clip to the bounding box}

\item{...}{further material properties passed to 3d plotting functions}
}
\description{
Draw 3d spherocylinders
}
\details{
The function requires the package \code{rgl} to be installed.
}
