% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolding.R
\name{parameters3d}
\alias{parameters3d}
\title{Original parameters}
\usage{
parameters3d(S)
}
\arguments{
\item{S}{list of spheroids}
}
\value{
list
}
\description{
Extract the original 3d size-shape-orientation parameters
}
\details{
The function simply extracts the parameters for the original 3d spheroid system
either for oblate or prolate spheroids and returns a list which consists of sizes
\code{a}, the orientation angles \code{Theta} and the shape parameters \code{s}.
}
