% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfoldr-package.R
\docType{package}
\name{unfoldr-package}
\alias{unfoldr-package}
\title{Stereological Unfolding for Spheroidal Particles}
\description{
Stereological unfolding as implemented in this package consists in
the estimation of the joint size-shape-orientation distribution of spheroidal
shaped particles based on the same measured quantities of corresponding planar
section profiles. A single trivariate discretized version of the (stereological)
integral equation in the case of prolate and oblate spheroids is solved
numerically by the EM algorithm. The estimation of diameter distribution of
spheres from planar sections (Wicksell's corpuscle problem) is also implemented.
Further, the package provides routines for the simulation of a Poisson germ-
grain process with either spheroids, spherocylinders or spheres as grains together
with functions for planar sections. For the purpose of exact simulation a bivariate size-shape
distribution is implemented.
}
\references{
\enumerate{
    \item Bene\eqn{\check{\textrm{s}}},
		  V. and Rataj, J. Stochastic Geometry: Selected Topics Kluwer Academic Publishers, Boston, 2004
	   	\item Ohser, J. and Schladitz, K. 3D images of materials structures Wiley-VCH, 2009
	    \item Ohser, J. and Muecklich, F. Statistical analysis of microstructures in materials science J. Wiley & Sons, 2000
     \item C. Lantu\eqn{\acute{\textrm{e}}}joul. Geostatistical simulation. Models and algorithms.
					Springer, Berlin, 2002. Zbl 0990.86007 
 }
}
